Revisit pipeline and use Shared Library
We could do another revision in the future if we determine that there is potential for the library to be enhanced or if we want to take advantage of Active vs Passive release cycle.

Definition of Active vs Passive release cycle
Active: There is a new release after any feature pushed to a protected branch (e.g. `master`)
Passive: Only release if there was a manual update to the repository to trigger a new release (e.g. updating package.json, or maybe adding a label to the PR in the future)

Signed-off-by: zFernand0 <fernando.rijocedeno@broadcom.com>