Update CODEOWNERS
I updated the file and comment out the unactive user and write the further steps to be followed by any admin. As i don't know the all project owners/maintainers so i leave it blank, this can be edited by the any owner later on. I request to any maintainer please update the file with your "@" in the 5th line of code to make you to review any PR created.

Signed-off-by: zaid <72139621+zaidmstrr@users.noreply.github.com>