fix: add support for record mode on data set download
does this by allowing record "encoding" to be specified in other encoding methods, rather than being in the top level quick pick which would be weird as it is not a common thing that would need to change. also, it cannot be true at the same time as binary mode, so by putting it into the encoding prompt as a supported option, it inherently bars both from being set at once. this could cause confusion however on how to specify record mode so a record mode item could be added to the top level quick pick instead in the future
Signed-off-by: JWaters02 <watersjoshua2002@gmail.com>